Use both old and new method for determining if "-hwe" Trello card should be used
A few Jenkins jobs [1][2][3] erroneously create new cards on Trello
instead of using the xenial-hwe card due to 9c5c9e64.  These jobs have
an `srutype` of `oem`.  We can fix this by combining the old and the new
method for determining if a "-hwe" Trello card should be used.

[1]: http://10.101.50.238:8080/job/cert-oem-sru-xenial-desktop-g7-7587-c1/
[2]: http://10.101.50.238:8080/job/cert-oem-sru-xenial-desktop-g7-7587-c2/
[3]: http://10.101.50.238:8080/job/cert-oem-sru-xenial-desktop-precision-5530-c6/